requirements.txt
dash
datasets
gensim
matplotlib
nltk
numpy<2.0.0
octis==1.13.1
pandas
plotly
pyarrow
scikit_learn
scipy==1.10.1
seaborn
sentence_transformers
setfit
setuptools
tqdm
umap_learn
wordcloud
community
networkx
python_louvain
pytorch_lightning
torch
langdetect
hdbscan
loguru
pre-commit